+++ spec/rubocop/cop/layout/indent_heredoc_spec.rb
@@ -3,6 +3,13 @@
 describe RuboCop::Cop::Layout::IndentHeredoc, :config do
   subject(:cop) { described_class.new(config) }
 
+  let(:allow_heredoc) { true }
+  let(:other_cops) do
+    {
+      'Metrics/LineLength' => { 'Max' => 5, 'AllowHeredoc' => allow_heredoc }
+    }
+  end
+
   shared_examples :offense do |name, code, correction = nil|
     it "registers an offense for #{name}" do
       inspect_source(code.strip_indent)
@@ -141,6 +148,26 @@ describe RuboCop::Cop::Layout::IndentHeredoc, :config do
 
         RUBY2
       RUBY
+
+      context 'when Metrics/LineLength is configured' do
+        let(:allow_heredoc) { false }
+
+        include_examples :offense, 'short heredoc', <<-RUBY, <<-CORRECTION
+          <<#{quote}RUBY2#{quote}
+          12
+          RUBY2
+        RUBY
+          <<#{quote}RUBY2#{quote}.strip_indent
+            12
+          RUBY2
+        CORRECTION
+
+        include_examples :accept, 'long heredoc', <<-RUBY
+          <<#{quote}RUBY2#{quote}
+          12345678
+          RUBY2
+        RUBY
+      end
 
       include_examples :check_message, 'suggestion powerpack',
                        [